Chore/#135: Presentation 그룹 정리
