--- Harubee-iOS/Sources/Helpers/Extensions/Bundle+.swift
@@ -1,17 +1,0 @@
-//
-//  BundleIdentifier.swift
-//  Shared
-//
-//  Created by 이정동 on 11/4/24.
-//  Copyright © 2024 namudiEtc. All rights reserved.
-//
-
-import Foundation
-
-
-extension Bundle {
-  static let app = Bundle(identifier: "etc.namudi.harubee-app")
-  static let domain = Bundle(identifier: "etc.namudi.harubee-domain")
-  static let data = Bundle(identifier: "etc.namudi.harubee-data")
-  static let shared = Bundle(identifier: "etc.namudi.harubee-shared")
-}
